requirements.txt
# Core dependencies
numpy>=1.24.0
torch>=2.0.0
pyyaml>=6.0

# Visualization
pygame>=2.5.0
matplotlib>=3.7.0

# Analysis and logging
pandas>=2.0.0
tqdm>=4.65.0

# Development
pytest>=7.3.0
black>=23.0.0
flake8>=6.0.0
mypy>=1.3.0
